+++ src/Registration.cpp
@@ -63,32 +63,33 @@ void registerAuxDataTypes()
     gtirb::AuxDataContainer::registerAuxDataType<PeResources>();
     gtirb::AuxDataContainer::registerAuxDataType<SouffleFacts>();
     gtirb::AuxDataContainer::registerAuxDataType<SouffleOutputs>();
+    gtirb::AuxDataContainer::registerAuxDataType<RawEntries>();
 }
 
 void registerDatalogLoaders()
 {
 #if defined(DDISASM_ARM_64)
-    // Register ELF-ARM64 target.
+    // Register ELF-ARM64-LE target.
     DatalogProgram::registerLoader(
         {gtirb::FileFormat::ELF, gtirb::ISA::ARM64, gtirb::ByteOrder::Little}, ElfArm64Loader);
 #endif
 
 #if defined(DDISASM_X86_32)
-    // Register ELF-X86 target.
+    // Register ELF-X86-LE target.
     DatalogProgram::registerLoader(
         {gtirb::FileFormat::ELF, gtirb::ISA::IA32, gtirb::ByteOrder::Little}, ElfX86Loader);
 
-    // Register PE-X86 target.
+    // Register PE-X86-LE target.
     DatalogProgram::registerLoader(
         {gtirb::FileFormat::PE, gtirb::ISA::IA32, gtirb::ByteOrder::Little}, PeX86Loader);
 #endif
 
 #if defined(DDISASM_X86_64)
-    // Register ELF-X64 target.
+    // Register ELF-X64-LE target.
     DatalogProgram::registerLoader(
         {gtirb::FileFormat::ELF, gtirb::ISA::X64, gtirb::ByteOrder::Little}, ElfX64Loader);
 
-    // Register PE-X64 target.
+    // Register PE-X64-LE target.
     DatalogProgram::registerLoader(
         {gtirb::FileFormat::PE, gtirb::ISA::X64, gtirb::ByteOrder::Little}, PeX64Loader);
 #endif